<fmt:formatNumber> 태그 발주detail 추가 및 합계 AMT 가 BIGINT 로 변경되어서 변수타입 long으로 변경, xml도 수정

+++ src/main/java/com/ess/erp/domain/OrderDetailDTO.java
@@ -11,5 +11,5 @@ public class OrderDetailDTO {
     private int unitPrice;      // UNIT_PRICE (단가)
     private int supplyAmt;      // SUPPLY_AMT (공급가액)
     private int vatAmt;         // VAT_AMT (부가세)
-    private int amt;            // AMT (합계금액)
+    private long amt;           // AMT (합계금액)
 }
